+++ src/react/index.ts
@@ -1,0 +1,4 @@
+export {connect} from './connect'
+export {connectAs} from './connectAs'
+export {connectToTree} from './connectToTree'
+export {connectToTreeAs} from './connectToTreeAs'
